requirements.txt
streamlit
pandas
python-binance
backtester
plotly
python-dotenv
python-binance
backtrader
plotly
backtrader_plotly

